+++ NEWS.md
@@ -1,3 +1,5 @@
+* Added `type = "age-scatter"` option for `autoplot.pop_data()`
+
 ## serocalculator 1.0.0
 
 * Moved underlying methods to `serocalculator` vignette